requirements.txt
python-vlc
youtube-search-python
pytube
